# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Revoked API
  version: "0.1.0"
  summary: Secret sharing and data collection over revocable, living grants.
  description: |
    Every operator runs this API under their own domain, so there is no shared
    base URL — substitute your own host everywhere below.

    Two things distinguish this API from a typical CRUD surface, and both are
    worth reading before integrating:

    - **Grants are living.** A share resolves the current value at read time,
      so rotating a secret updates every share that points at it. Pausing or
      revoking one takes effect on the next read, not on the next sync.
    - **Public endpoints are protected by slug entropy alone.** Anything under
      `/api/public` or `/s` is reachable by whoever holds the slug. There is
      deliberately no list endpoint for them.

    Endpoints marked *unauthenticated* are meant to be called by people with no
    account on this server — recipients of a share, responders to a request,
    verifiers walking the trust chain.
  license:
    name: See repository
servers:
  - url: https://{host}
    description: Your deployment
    variables:
      host:
        default: api.example.com
        description: The domain this operator runs under, matching its DNS TXT record.
security:
  - ApiKey: []
  - Session: []
...
paths:
...
  /api/connections/authorize:
    post:
      tags: [Connections]
      summary: Authorize a tool (from the app)
      description: |
        Called by the owner's app after they agreed to connect a tool. Finds or
        creates the connection for the tool's origin in the active workspace,
        renews it for 90 days (tokens of a connection that had lapsed are
        deleted, not revived), and returns a one-time code, valid for five minutes, bound to the tool's PKCE
        challenge and return address. The app then opens the return address with
        `code` and the tool's `state`.
      security:
        - Session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId, redirectUri, challenge]
              properties:
                clientId: { type: string, example: "https://mietunterlagen.revoked.link", description: "The tool's origin." }
                clientName: { type: string, maxLength: 40, description: "The name the tool gives itself; shown as a claim." }
                redirectUri: { type: string, description: "Where the code is delivered; must be on `clientId`'s origin." }
                allowRevoke:
                  type: boolean
                  description: |
                    Whether the tool may revoke the links its proposals became.
                    Left out, a connection keeps what it had; a new one starts
                    without it.
                allowHandOver:
                  type: boolean
                  description: |
                    Whether the tool may receive links the owner shares with
                    it. Left out, a connection keeps what it had; a new one
                    starts without it. A link can only be created with
                    `handedOver` while this is set
                    (`link_hand_over_not_allowed` otherwise), and the tool
                    stops seeing the URLs it was given when it is taken away.
                challenge: { type: string, description: "PKCE S256 challenge, base64url." }
                reuse:
                  type: boolean
                  description: |
                    Set when the owner was not asked again because the tool is
                    already connected and another of their browsers wants in.
                    Only issues a code: the connection keeps its name,
                    permissions and expiry (`allowRevoke` and `allowHandOver`
                    are ignored). Refused with `connection_not_connected` when
                    there is no live connection for the origin.
                poll:
                  type: boolean
                  description: |
                    Set when the page that asked collects the answer itself:
                    the tool named the owner's server in its link (`poll`), so
                    the app opens no return address and the page exchanges its
                    verifier alone at `/api/connect/token`. No return address
                    ties such an answer to one of the owner's own browsers, so
                    the app never sets it unasked, `reuse` included: the owner
                    compares a code both ends derive from the challenge (the
                    first three bytes of its SHA-256, as `83C-047`).
      responses:
        "200":
          description: The one-time code.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
                  connectionId: { type: string }
        "400":
          description: "`connection_client_invalid`, `connection_redirect_invalid` or `connection_challenge_invalid`."
        "409":
          description: "`connection_not_connected`: `reuse` was set, but the tool is not connected or its connection has lapsed."
  /api/connect/token:
    post:
      tags: [Connections]
      summary: Exchange a code for a connection token (the tool)
      description: |
        Spends the one-time code — right or wrong, a code works once — and, when
        the verifier matches its challenge and the return address matches,
        returns a token for this browser. Rate-limited per IP.

        Without `code`, the page asks for an answer the app left for it
        (`poll` on authorize): it is found by the verifier's challenge, and
        given only to a request whose `Origin` is the tool's own. Until the
        owner has answered there is nothing to find, and the reply is `202`;
        a tool asks again every few seconds, for as long as an answer is kept
        (five minutes). An answer issued as a code is never found this way.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verifier, redirectUri]
              properties:
                code: { type: string, description: "Left out by a page that collects its answer itself." }
                verifier: { type: string, description: "The PKCE verifier." }
                redirectUri: { type: string, description: "The same return address the code was issued for." }
      responses:
        "200":
          description: The token, shown once.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  connection:
                    type: object
                    properties:
                      id: { type: string }
                      clientName: { type: string }
                      expiresAt: { type: string, description: "When the connection lapses." }
        "202":
          description: No code was sent, and the owner has not answered yet.
          content:
            application/json:
              schema:
                type: object
                properties:
                  pending: { type: boolean }
        "400":
          description: "`connection_grant_invalid`."
        "429":
          description: "`rate_limited`."
  /api/connection:
    get:
      tags: [Connections]
      summary: The status of the tool's links
      description: |
        The links the tool's proposals became — status, views, expiry. `url`
        appears only on links the owner handed to the tool, and only while they
        can be opened. A tool stores nothing here: `ref` and `label` are what it
        set on the proposal.
      security:
        - Connection: []
      responses:
        "200":
          description: The connection as the tool may see it.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  clientId: { type: string }
                  clientName: { type: string }
                  expiresAt:
                    type: string
                    description: |
                      When the connection lapses, 90 days after the owner last
                      connected the tool. Only connecting again from the app
                      renews it.
                  allowRevoke:
                    type: boolean
                    description: Whether the owner lets the tool revoke its links.
                  allowHandOver:
                    type: boolean
                    description: Whether the owner lets the tool receive links.
                  links:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        ref: { type: string, description: "The tool's own reference from the proposal." }
                        label: { type: string }
                        status: { type: string, enum: [active, paused, revoked, expired] }
                        viewCount: { type: integer }
                        maxViews: { type: integer }
                        expiresAt: { type: string }
                        created: { type: string }
                        handedOver: { type: boolean }
                        url: { type: string, description: "Only when handed over and active." }
        "401":
          description: "`connection_unauthorized`, or `connection_expired` once the connection has lapsed."
    delete:
      tags: [Connections]
      summary: Disconnect (the tool)
      description: Deletes the connection. The owner's links stay; they no longer name the tool.
      security:
        - Connection: []
      responses:
        "204":
          description: Disconnected.
        "401":
          description: "`connection_unauthorized`, or `connection_expired` once the connection has lapsed."
  /api/connection/links/{id}/revoke:
    post:
      tags: [Connections]
      summary: Revoke one of the tool's links (the tool)
      description: |
        Ends a link the tool's proposal became. Only when the owner allowed it
        (`allowRevoke`); a tool can end a share, never start one or read it.
        The owner is notified as for any revocation. Revoking a link that is
        already revoked succeeds.
      security:
        - Connection: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The link as the tool may see it, now revoked.
        "401":
          description: "`connection_unauthorized` or `connection_expired`."
        "403":
          description: "`connection_revoke_not_allowed`."
        "404":
          description: Not a link of this tool.
  /api/connections/{id}/permissions:
    post:
      tags: [Connections]
      summary: Change what a tool is allowed (the owner)
      description: |
        The two things the owner decides per connection. A field left out
        stays as it is.
      security:
        - Session: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                allowRevoke: { type: boolean, description: "The tool may revoke the links its proposals became." }
                allowHandOver: { type: boolean, description: "The tool may receive the links its proposals became." }
      responses:
        "200":
          description: Both permissions as they now stand.
        "404":
          description: Not one of your connections.
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Programmatic access. Sending a key as a Bearer token authenticates as
        nobody — the request is treated as a guest and refused by the rules.
    Session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A session token from sign-in. Used by the app.
    Connection:
      type: apiKey
      in: header
      name: X-Revoked-Connection
      description: A connected tool's token, from `/api/connect/token`.
...